src/MedEquity.ML/requirements.txt
grpcio==1.73.1
grpcio-tools==1.73.1
protobuf==6.31.1
google-genai==1.16.1
